!!! FEATURE: Use cldr format for dates
With this change the cldr format is
used to define the date format.

This is breaking as the previously used
dateFormat setting will not do anything anymore.

If the format doesn’t match the requirement,
the cldrFormat setting needs to be set according
to http://cldr.unicode.org/translation/date-time-patterns.